♻️ Move MAX_POLL_OPTIONS to poll feature constants
